- install.rdf now handles the version passed to packages.sh - minor script improvement in packages.sh - removed warning with setTimeout(), now uses nsITimer() intead. - fixed timer loop that prevented the modification of submitted title   and artist name to lyricsfly.com - prevent repetitive exception when the lyrics window is submitted/closed

+++ branches/2.0_no_db/package.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+
+if [ -z "$1" ]
+then
+	echo "$0 release_number"
+	exit 1
+fi
+
 cd ..
 rm -rf mpm_release
 svn export 2.0_no_db mpm_release
